@@ -2,12 +2,23 @@ import { Location, PopStateEvent } from '@angular/common';
 import {
   Component,
   HostListener,
+  inject,
   Inject,
   LOCALE_ID,
-  OnInit
+  OnInit,
+  ViewChild,
+  ViewContainerRef
 } from '@angular/core';
 import { FormBuilder, FormControl, FormGroup } from '@angular/forms';
 import { ActivatedRoute, Params, Router } from '@angular/router';
+import {
+  MaintenanceItem,
+  MaintenanceScheduleService,
+  MaintenanceSettings
+} from '@europeana/metis-ui-maintenance-utils';
+
+import { cookieConsentConfig } from '../environments/eu-cm-settings';
+import { maintenanceSettings } from '../environments/maintenance-settings';
 import { SubscriptionManager } from './subscription-manager';
 import { APIService, ClickService } from './_services';
 import {
@@ -18,11 +29,15 @@ import {
 } from './_models';
 import { LandingComponent } from './landing';
 import { OverviewComponent } from './overview';
+
 @Component({
   selector: 'app-root',
+  styleUrls: ['./app.component.scss'],
   templateUrl: './app.component.html'
 })
 export class AppComponent extends SubscriptionManager implements OnInit {
+  private readonly maintenanceService = inject(MaintenanceScheduleService);
+
   formCTZero: FormGroup<{ contentTierZero: FormControl<boolean> }>;
   landingData: GeneralResultsFormatted;
   landingComponentRef: LandingComponent;
@@ -30,6 +45,10 @@ export class AppComponent extends SubscriptionManager implements OnInit {
   showPageTitle: boolean;
   lastSetContentTierZeroValue = false;
   skipLocationUpdate = false;
+  maintenanceInfo?: MaintenanceItem = undefined;
+
+  @ViewChild('consentContainer', { read: ViewContainerRef })
+  consentContainer: ViewContainerRef;
 
   constructor(
     private readonly api: APIService,
@@ -41,6 +60,22 @@ export class AppComponent extends SubscriptionManager implements OnInit {
   ) {
     super();
     document.title = 'Statistics Dashboard';
+    this.checkIfMaintenanceDue(maintenanceSettings);
+    this.showCookieConsent();
+  }
+
+  checkIfMaintenanceDue(settings: MaintenanceSettings): void {
+    this.maintenanceService.setApiSettings(settings);
+    this.subs.push(
+      this.maintenanceService
+        .loadMaintenanceItem()
+        .subscribe((item: MaintenanceItem | undefined) => {
+          this.maintenanceInfo = item;
+          if (item?.maintenanceMessage && this.landingComponentRef) {
+            this.landingComponentRef.isLoading = false;
+          }
+        })
+    );
   }
 
   /** buildForm
@@ -185,6 +220,33 @@ export class AppComponent extends SubscriptionManager implements OnInit {
       component.locale = this.locale;
       this.landingComponentRef = undefined;
       this.showPageTitle = false;
+    }
+  }
+
+  /**
+   * showCookieConsent
+   * - calls show on cookieConsent
+   **/
+  async showCookieConsent(force = false): Promise<void> {
+    const CookieConsentComponent = (
+      await import('@europeana/metis-ui-consent-management')
+    ).CookieConsentComponent;
+
+    this.consentContainer.clear();
+
+    const cookieConsent = this.consentContainer.createComponent(
+      CookieConsentComponent
+    );
+
+    cookieConsent.setInput('translations', cookieConsentConfig.translations);
+    cookieConsent.setInput('services', cookieConsentConfig.services);
+    cookieConsent.setInput('fnLinkClick', (): void => {
+      cookieConsent.instance.shrink();
+      this.router.navigate(['/cookie-policy']);
+    });
+
+    if (force) {
+      cookieConsent.instance.show();
     }
   }
 
